# Introduce PureComponent to reduce `render` method calling