--- a/docs/api/Ovto/State/MissingValue.html
+++ b/docs/api/Ovto/State/MissingValue.html
@@ -125,9 +125,9 @@
 </div>
 
       <div id="footer">
-  Generated on Sat Apr 20 20:59:20 2019 by
+  Generated on Mon Jul 22 23:03:46 2019 by
   <a href="http://yardoc.org" title="Yay! A Ruby Documentation Tool" target="_parent">yard</a>
-  0.9.19 (ruby-2.5.5).
+  0.9.19 (ruby-2.7.0).
 </div>
 
     </div>