# What is PureComponent?


PureComponent is a Component, but it caches `render`'s result to skip `render` method calling if the attributes are same.
I got the idea from React's PureComponent.


Example

```ruby
class Hello < Ovto::PureComponent
  def render(name:)
    o 'h1', "Hello, #{compute_name_with_heavy_process(name)}"
  end

  private def compute_name_with_heavy_process(name)
    name.do_some_heavy_process
  end
end


class MainComponent < Ovto::Component
  def render
    o 'div' do
      o Hello, name: state.name
      o 'div', state.something
    end
  end
end
```



In this example, `Hello#render` has a heavy function. But it is only called when `state.name` is changed because it uses PureComponent. It is not called when `state.something` is changed.
So it will be faster than without PureComponent.









## Limitation


PureComponent only supports component that does not use `state`.
So `state` method is not available for PureComponent.


PureComponent decides using cache by attributes changes.
If it is aware of state, it will be meaningless. Because state is always changed when `render` is called.



But I think we can support state with the manual specifying state, which is like shouldUpdateComponent of React.
For example

```ruby
class C < Ovto::Component
  def render
    o 'h1', state.name
  end

  def should_update?(attr, next_attr, state, next_state)
    # List dependent states manually
    state.name != next_state.name
  end
end
```

But I do not like this idea. It needs listing all of the dependent states manually. If I forget a dependency, it will cause a mysterious bug.

And we can add the `should_update?` feature in the future if it is needed.
So I think this feature is not necessary for the first release of PureComponent.





## Benchmarking

I profiled my application that is based on Ovto with/without PureComponent.
https://puzzle.pocke.me/

It is a Number Place (a.k.a. Sudoku) application. And I profiled clicking a cell.
When clicking a cell, it calls `render` method only for the clicked cell if the cell component is a PureComponent.



With PureComponent
![190719220922](https://user-images.githubusercontent.com/4361134/61537420-1c577c80-aa72-11e9-863c-ed48c4d954db.png)



Without PureComponent
![190719221017](https://user-images.githubusercontent.com/4361134/61537421-1c577c80-aa72-11e9-9dbf-bb3778771e48.png)


As you can see, the second peek is the `render` method calling. (The first peek is handling the click event.)
The render time is about 29ms (765ms ~ 794ms) without PureComponent, but it is about 5.5ms (692ms ~ 697.5ms).
It reduces 23.5ms, it is 5.3x faster.


I don't think 23.5ms is a meaningful difference. I cannot distinguish the difference.
But I think it will be a meaningful difference if the application will be larger, or the user uses a low-tech computer, such as a cheap smartphone.
If the difference is 4~5x, which is 100ms, I guess I will be annoyed by the delay.



## Implementation

This pull request will change the Component (NOT PureComponent) implementation.
It will cache child components instances by this change because component instance should remember the previous attributes.



By the way, the PureComponent implementation is super simple. See the code.






---


What do you think? If you accept this proposal, I'll do the following TODOs.


- [x] Write test
- [x] Write documentation